Fix newer Telegram ID support and add setup documentation
Telegram user IDs now exceed the 32-bit integer limit (~2.1B).
All chat_id columns in User, Group, Bot, and Message models changed
from IntegerField to BigIntegerField so newer accounts work.

- Add migration script (botlistbot/migration/bigint_ids.py) for
  existing databases to widen INTEGER columns to BIGINT
- Rewrite README with a copy-paste quickstart guide
- Clarify template.env with sections and inline comments
- Cast DEVELOPER_ID to int in settings.py

+++ botlistbot/models/bot.py
@@ -51,7 +51,7 @@ class Bot(BaseModel):
 
     userbot = BooleanField(default=False)
     botbuilder = BooleanField(default=False)
-    chat_id = IntegerField(null=True)
+    chat_id = BigIntegerField(null=True)
 
     approved = BooleanField(default=True)
     submitted_by = ForeignKeyField(User, null=True, related_name='submitted_by')
